Added: - all the necessary calculations from mean to sharp Ratio

+++ src/invest/data/model/dailysellhistory/DailySellHistoryMapper.kt
@@ -10,7 +10,7 @@ object DailySellHistoryMapper {
 
         val newPrice = dailySellHistoryDto.price
 
-        val newVolume = dailySellHistoryDto.volume.toInt()
+        val newVolume = dailySellHistoryDto.volume
 
         return DailySellHistory(
             date = newDate,
